requirements.txt
Pillow>=9.0.0
python-dateutil>=2.8.2
PyQt6>=6.6.0
pytest-qt>=5.0.0

